+++ Netable/NetableExample/View Controllers/EmptyLoggerViewController.swift
@@ -12,12 +12,11 @@ import UIKit
 class EmptyLoggerViewController: UIViewController {
     @IBOutlet weak var imageView: UIImageView!
 
-    /// Create a Netable instance that won't record any logs
+    /// Pass in an instance of `EmptyLogDestination` to have Netable not print any logs.
      private let netable = Netable(baseURL: URL(string: "https://api.thecatapi.com/v1/")!, logDestination: EmptyLogDestination())
 
      override func viewDidLoad() {
          super.viewDidLoad()
-         // Do any additional setup after loading the view.
 
          netable.request(GetCatRequest()) { [weak self] result in
              guard let self = self else { return }