+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added a new post example
 - Added ability to customize the decode method for requests
+- Update file headers, add some more comments to examples
 
 ### Changed
 - Reworked the example project to allow for more examples to be added